vsg-models/vsg-gen/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vsg-gen"
version = "0.1.0"
description = "Video Snippets Generator"
authors = ["Ricardo G. Sousa <ricardo@rsousa.co>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
llama-cpp-python = "~0.2.57"


